requirements.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==3.7.0
django-crontab==0.7.1
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
geographiclib==1.50
geopy==2.1.0
gunicorn==20.1.0
phonenumbers==8.12.23
Pillow==8.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
